fix load when not logined
