+++ lib/pages/articlepage.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:project_finaltask/components/appbar.dart';
 import 'package:project_finaltask/models/article.dart';
