+++ lib/pages/followspage.dart
@@ -1,0 +1,1 @@
+//カード
