+++ lib/secret.dart
@@ -1,0 +1,5 @@
+final clientID = '806515abce77408e98feed9ab93cab135a43182f';
+final clientSecret = '0c4fe0a00f290f04f905f6a982b2ed4848b615d6';
+final keyAccessToken = 'qiita/accessToken';
+
+// 'df9586e167ea06de68a460858b20b6a237d2b3a2'; //登録したアクセストークン
