+++ lib/views/articlelistview.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:project_finaltask/pages/articlepage.dart';
 import 'package:project_finaltask/utils/changedate.dart';
