+++ pubspec.yaml
@@ -48,6 +48,7 @@ dev_dependencies:
 flutter:
   assets:
     - assets/images/backgroundimage.png
+    - assets/images/networkimage.png
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
